Make TodWindowSpeaker delete channel contents only when *not* in debug mode.
The current debug setup for Discord interactions doesn't play with this well.

+++ src/main/java/com/edmazur/eqrs/discord/speaker/TodWindowSpeaker.java
@@ -1,5 +1,6 @@
 package com.edmazur.eqrs.discord.speaker;
 
+import com.edmazur.eqrs.Config;
 import com.edmazur.eqrs.discord.Discord;
 import com.edmazur.eqrs.discord.DiscordChannel;
 import com.edmazur.eqrs.discord.DiscordPredicate;
@@ -11,14 +12,17 @@ public class TodWindowSpeaker implements Runnable {
 
   private static final DiscordChannel CHANNEL = DiscordChannel.TIMERS;
 
+  private final Config config;
   private final Discord discord;
   private final RaidTargetTableMaker raidTargetTableMaker;
   private final DiscordTableFormatter discordTableFormatter;
 
   public TodWindowSpeaker(
+      Config config,
       Discord discord,
       RaidTargetTableMaker raidTargetTableMaker,
       DiscordTableFormatter discordTableFormatter) {
+    this.config = config;
     this.discord = discord;
     this.raidTargetTableMaker = raidTargetTableMaker;
     this.discordTableFormatter = discordTableFormatter;
@@ -28,7 +32,12 @@ public class TodWindowSpeaker implements Runnable {
   public void run() {
     try {
       Table table = raidTargetTableMaker.make();
-      discord.deleteMessagesMatchingPredicate(CHANNEL, DiscordPredicate.isFromYourself());
+      // Delete channel contents only when *not* in debug mode. The current debug setup for Discord
+      // interactions doesn't play with this well.
+      // TODO: Remove this restriction once a better test setup is in place.
+      if (!config.getBoolean(Config.Property.DEBUG)) {
+        discord.deleteMessagesMatchingPredicate(CHANNEL, DiscordPredicate.isFromYourself());
+      }
       for (String text : discordTableFormatter.format(table)) {
         discord.sendMessage(CHANNEL, text);
       }
